Move task subclasses into the task package
