+++ core/dsl/src/main/kotlin/edu/wpi/axon/dsl/task/ConstructYoloV3ImageInput.kt
@@ -1,8 +1,8 @@
-package edu.wpi.axon.dsl.variable
+package edu.wpi.axon.dsl.task
 
 import edu.wpi.axon.dsl.Code
 import edu.wpi.axon.dsl.Import
-import edu.wpi.axon.dsl.task.Task
+import edu.wpi.axon.dsl.variable.Variable
 
 /**
  * Constructs the input argument for the YoloV3 model.
